Relocated env data endpoint unit tests
