+++ src/tests/tests.py
@@ -28,177 +28,7 @@ class MainTests(TestCase):
         client = Client()
         response = client.get('/tests/')
         self.assertEqual(response.status_code, 200)
-
-    def test_response_temperature_data(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'temperature',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 200)
-
-    @patch('cassy.get_env_data')
-    def test_response_env_data_exception(self, env_data_mock):
-        '''
-        Test env_data endpoint when get_env_data throws Exception
-        Args:
-            env_data_mock:
-
-        Returns:
-
-        '''
-        setup_test_environment()
-        client = Client()
-        env_data_mock.side_effect = Exception('Test exception')
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'temperature',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('env_data_unknown' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_humidity_data(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'humidity',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 200)
-
-    def test_response_leafwetness_data(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'leafwetness',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 200)
-
-    def test_response_invalid_vineyard(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '-1',
-            'env_variable': 'temperature',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_id_not_found' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_invalid_env_variable(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'cheesiness',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('env_data_invalid' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_invalid_env_data_request_missing_vineyard(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'env_variable': 'temperature',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_no_id' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_invalid_env_data_request_missing_env_variable(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('env_data_invalid' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_invalid_env_data_request_non_integer_id(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': 'asdf',
-            'env_variable': 'temperature',
-            'auth_token': os.environ.get('LOGIN_SEC_TOKEN')
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        error = json.loads(response.content.decode('utf-8'))['errors']
-        self.assertTrue('vineyard_bad_id' in error)
-        self.assertEqual(response.status_code, 400)
-
-    def test_response_env_data_invalid_token(self):
-        setup_test_environment()
-        client = Client()
-        body = {
-            'vineyard_id': '0',
-            'env_variable': 'temperature',
-            'auth_token': 'chestercheetah'
-        }
-        response = client.post(
-            '/env_data',
-            data=json.dumps(body),
-            content_type='application/json'
-        )
-        self.assertEqual(response.status_code, 403)
-
+        
     def test_admin_redirect(self):
         setup_test_environment()
         client = Client()
